more Darwin warning fixes

+++ crserverlib/get_sizes.py
@@ -330,7 +330,7 @@ static unsigned int __numValues( GLenum pname )
 		if (temp->pname == pname)
 			return temp->num_values;
 	}
-	crError( "Invalid pname to __numValues: 0x%x\\n", pname );
+	crError( "Invalid pname to __numValues: 0x%x\\n", (int) pname );
 	return 0;
 }
 """
